Initial Solidity migration (#279)
* WIP migration

* Migration WIP

* Compiling

* Test fix

* Fix address references

* Test fix

* Test fix

* .skip

* Update package.json

+++ contracts/interfaces/ERC1404.sol
@@ -18,7 +18,7 @@ interface ERC1404
   function detectTransferRestriction(
     address from,
     address to,
-    uint256 value
+    uint value
   ) external view
     returns (uint8);
